Move prowjob extract to pkg extractimages, fix test error message
